add $keep property to protocol definitions
This commit adds the ability to include a "$keep" property within a protocol definition structure path.

The goal of this is to allow for users to limit to the number of records under a specific path.
If that number is reached, the newest messages are kept and the older ones beyond the limit are purged.